// app/types/types.tsx
export type services_3 = {
    id: number;
title: string;
description: string;
related_subcategory_2: number;
price: number;
order_type: "cart" | "apply-form" | "cart_with_units";
image_url: string;
}


export type addOn = {
  id: number;
  title: string;
  description: string;
  price: number;
  related_service_3: number;
  image_url: string 
}